requirements/dev.txt
black==23.12.1
isort==5.13.2
flake8==7.0.0
pre-commit==3.6.0
pytest-asyncio==0.23.3
httpx==0.26.0